remove images_path when extracting keypoints

+++ action_similarity/motion.py
@@ -189,11 +189,13 @@ def extract_keypoints(video_path: str, fps: int) -> Dict[int, Dict]:
 
     video_name, _ = os.path.splitext(video_path)
     images_path = os.path.join(video_name, 'images')
-    shutil.rmtree(images_path)
+    if os.path.exists(images_path):
+        shutil.rmtree(images_path)
     os.makedirs(images_path, exist_ok=True)
     
     json_path = os.path.join(video_name, 'json')
-    shutil.rmtree(json_path)
+    if os.path.exists(json_path):
+        shutil.rmtree(json_path)
     os.makedirs(json_path, exist_ok=True)
 
     clip = mpy.VideoFileClip(video_path)
